refactor(mentions): move the shared mention pieces out of comments
The mention overlay, its suggestion model and the two-tier search helper
are all generic, but they live under `blocs/comments/` and
`screens/comments/`. The caption composer needs the same three, and
importing them from the comments feature would point a metadata widget
at a comments screen.

Moves them to `lib/mentions/` and `lib/widgets/mentions/`, with the
overlay test following to its mirror path. `MentionSuggestion` was
declared inside the composer bloc's `part of` state file and is now a
standalone class the bloc imports.

No behaviour change: the widget, the search tiers and the model are
unchanged, only relocated.
