+++ mobile/lib/blocs/comments/comment_composer/comment_composer_bloc.dart
@@ -4,9 +4,10 @@ import 'package:bloc_concurrency/bloc_concurrency.dart';
 import 'package:comments_repository/comments_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:openvine/blocs/comments/comment_composer/mention_search.dart';
 import 'package:openvine/blocs/comments/comment_composer/reportable_sites.dart';
 import 'package:openvine/blocs/comments/comments_list/comments_list_helpers.dart';
+import 'package:openvine/mentions/mention_search.dart';
+import 'package:openvine/mentions/mention_suggestion.dart';
 import 'package:openvine/observability/reportable_error.dart';
 import 'package:openvine/services/auth_service.dart';
 import 'package:openvine/services/mention_resolution_service.dart';
