+++ mobile/lib/screens/comments/widgets/comment_input.dart
@@ -3,10 +3,10 @@
 
 import 'package:divine_ui/divine_ui.dart';
 import 'package:flutter/material.dart';
-import 'package:openvine/blocs/comments/comment_composer/comment_composer_bloc.dart';
 import 'package:openvine/constants/text_scale_limits.dart';
 import 'package:openvine/l10n/l10n.dart';
-import 'package:openvine/screens/comments/widgets/mention_overlay.dart';
+import 'package:openvine/mentions/mention_suggestion.dart';
+import 'package:openvine/widgets/mentions/mention_overlay.dart';
 
 /// Input widget for posting new top-level comments.
 ///
