+++ mobile/lib/screens/comments/widgets/widgets.dart
@@ -1,13 +1,14 @@
 // ABOUTME: Barrel file exporting all comments widgets
 // ABOUTME: Provides single import point for comment-related UI components
 
+export 'package:openvine/widgets/mentions/mention_overlay.dart';
+
 export 'comment_input.dart';
 export 'comment_item.dart';
 export 'comment_options_modal.dart';
 export 'comment_skeleton_loader.dart';
 export 'comments_empty_state.dart';
 export 'comments_list.dart';
-export 'mention_overlay.dart';
 export 'new_comments_pill.dart';
 export 'pending_video_reply_tile.dart';
 export 'unfocus_on_sheet_dismiss.dart';
