+++ mobile/test/blocs/comments/comment_composer/comment_composer_bloc_test.dart
@@ -9,7 +9,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:models/models.dart' show UserProfile;
 import 'package:openvine/blocs/comments/comment_composer/comment_composer_bloc.dart';
-import 'package:openvine/blocs/comments/comment_composer/mention_search.dart';
+import 'package:openvine/mentions/mention_search.dart';
+import 'package:openvine/mentions/mention_suggestion.dart';
 import 'package:openvine/services/auth_service.dart' hide UserProfile;
 import 'package:openvine/services/mention_resolution_service.dart';
 import 'package:profile_repository/profile_repository.dart';
