+++ mobile/test/screens/comments/comment_input_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:openvine/blocs/comments/comment_composer/comment_composer_bloc.dart';
 import 'package:openvine/l10n/generated/app_localizations.dart';
+import 'package:openvine/mentions/mention_suggestion.dart';
 import 'package:openvine/screens/comments/comments.dart';
 
 Finder _divineIcon(DivineIconName name) =>
