+++ mobile/test/widgets/mentions/mention_overlay_test.dart
@@ -4,10 +4,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:openvine/blocs/comments/comment_composer/comment_composer_bloc.dart';
 import 'package:openvine/l10n/l10n.dart';
-import 'package:openvine/screens/comments/widgets/mention_overlay.dart';
+import 'package:openvine/mentions/mention_suggestion.dart';
 import 'package:openvine/services/nip05_verification_service.dart';
+import 'package:openvine/widgets/mentions/mention_overlay.dart';
 
 void main() {
   const pubkey =
